Deployment/requirements.txt

tensorflow
gradio
pillow
